% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dpred}
\alias{dpred}
\title{Table of 372 SMILES with predicted internal energy E and HOMO-LUMO gap}
\format{A data.frame with 372 rows of 2 variables:
\describe{
  \item{E}{internal energy, in kJ/mol}
  \item{HOMO-LUMO gap}{HOMO-LUMO gap, in eV}
  ...
}}
\usage{
dpred
}
\description{
A dataset containing the predicted internal energy E and HOMO-LUMO gap of
372 generated compounds with iqspr
}
\keyword{datasets}
