% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENgram.R
\docType{class}
\name{ENgram-class}
\alias{ENgram-class}
\alias{ENgram}
\title{Extended N-gram model for learning SMILES strings}
\description{
Extended N-gram model for learning SMILES strings
}
\section{Methods}{

\describe{
\item{\code{initialize(smis = NULL, order = NULL)}}{Initialize the extend N-gram model with SMILES strings smis and numeric value order}

\item{\code{update_mat(data)}}{update the model with additional SMILES strings data}
}}

\examples{
data(trainedSMI)
data(engram_5k)  #same as run => engram <- ENgram$new(trainedSMI, order=10)
#-----arguments
#smis: character vector, SMILES string set for training an extended N-gram model
#order: numeric, value representing the maximum order in modified N-gram model
#-----arguments

}
