% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{engram_5k}
\alias{engram_5k}
\title{n-grams model learned from 5,000 SMILES from PubChem}
\format{A ENgram object only suitable for the iqspr
\describe{
  ...
}}
\source{
\url{https://pubchem.ncbi.nlm.nih.gov/}
}
\usage{
engram_5k
}
\description{
A dataset containing up to order 10 n-grams issued from 5,000 SMILES
}
\keyword{datasets}
