% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gensmis}
\alias{gensmis}
\title{Table of 372 SMILES with associated QSPRScore}
\format{A matrix with 372 rows of 2 variables:
\describe{
  \item{SMILES}{SMILES, in string of characters}
  \item{QSPRScore}{generation score, in a.u.}
  ...
}}
\usage{
gensmis
}
\description{
A dataset containing 372 generated SMILES with their QSPRScore
}
\keyword{datasets}
