% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptor.R
\name{get_descriptors}
\alias{get_descriptors}
\title{get a list of the available descriptors types}
\usage{
get_descriptors()
}
\value{
a list of the available descriptors types
}
\description{
get a list of the available descriptors types
}
\examples{
list_descriptors <- get_descriptors()
}
