% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIModels.R
\name{get_Model_params}
\alias{get_Model_params}
\title{get a list of default parameters for a given regression algorithm}
\usage{
get_Model_params(model_name = "elasticnet")
}
\arguments{
\item{model_name}{is the model's name for which you want to display the default parameters ("elasticnet" by default).
For the "svm" model, a.k.a Support Vector Machine, Gaussian RBF or Laplace kernel ("rbfdot" or "laplacedot" respectively)
are supported (limitation from the \code{\link[kernlab]{sigest}} function used as optimizer on the sigma parameter).}
}
\value{
a list of the default parameters for a given regression algorithm.
}
\description{
Get a list of default parameters for a given regression algorithm.
}
\examples{
list_params <- get_Model_params(model_name = "elasticnet")
}
