% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIModels.R
\name{get_Models}
\alias{get_Models}
\title{get a list of the implemented regression algorithms}
\usage{
get_Models()
}
\value{
a list of the implemented regression algorithms with their nickname, as implemented in iqspr, and their common
name from literature.
}
\description{
Get a list of the implemented regression algorithms.
}
\examples{
list_models <- get_Models()
}
