% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{predict properties with forward QSPR model}
\usage{
predict(qsprpred, smis)
}
\arguments{
\item{qsprpred}{QSPRpred class object}

\item{smis}{SMILES strings}
}
\description{
predict properties with forward QSPR model
}
\examples{
data(qsprpred_EG_5k)
predict(qsprpred_EG_5k, c("c1ccccc1O", "c1ccccc1F"))

}

