% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predictions}
\alias{predictions}
\title{List of predictions associated to the internal energy E and HOMO-LUMO gap
of 500 SMILES}
\format{A list with two matrices of 500 rows of 2 variables:
\describe{
  ...
}}
\usage{
predictions
}
\description{
A list containing two matrices of predicted internal energy E and HOMO-LUMO gap
for 500 SMILES from linear and random forest models respectively
}
\keyword{datasets}
