% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predinit}
\alias{predinit}
\title{Table of predicted internal energy E and HOMO-LUMO gap for the phenol}
\format{A matrix with 1 row of 2 variables:
\describe{
  \item{E}{internal energy, in kJ/mol}
  \item{HOMO-LUMO gap}{HOMO-LUMO gap, in eV}
  ...
}}
\usage{
predinit
}
\description{
A dataset containing the predicted internal energy E and HOMO-LUMO gap
for the phenol
}
\keyword{datasets}
