\name{qsprpred_EG_5k}
\alias{qsprpred_EG_5k}
\docType{data}
\title{
QSPRpred object trained with 5000 SMILES strings and properties (Internal energy and HOMO-LUMO gap) dataset
}
\description{
QSPRpred object trained with 5000 SMILES strings and properties (Internal energy and HOMO-LUMO gap) dataset
}
\usage{data("engram_5k")}
\examples{
data(engram_5k)
}
\keyword{datasets}
