\name{trainedSMI}
\alias{trainedSMI}
\docType{data}
\title{
sample dataset for learning with Extended N-gram model
}
\description{
sample dataset for learning with Extended N-gram model
5000 SMILES strings are included.
}
\usage{data("qspr.data")}
\examples{
data(trainedSMI)
trainedSMI[1:10]
}
\keyword{datasets}
