% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewstr.R
\name{viewstr}
\alias{viewstr}
\title{view 2D structures from SMILES string vector}
\usage{
viewstr(smis)
}
\arguments{
\item{smis}{SMILES string vector}
}
\description{
view 2D structures from SMILES string vector
}
\examples{
viewstr(c("c1ccc2ccc3c(NCCN(C)C)cc(nc3c2c1)", "c1ccc2ccc3c(NCCN(CC)CCCl)cc(nc3c2c1)",
 "c1ccc2ccc3c(NC(CC)CC)cc(nc3c2c1)", "c1ccc2ccc3c(c2c1)ncc(c3NCCNCC=CCCCC)"))

}
