% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qspr.data}
\alias{qspr.data}
\title{Table of 16,674 SMILES with associated internal energy E and HOMO-LUMO gap}
\format{A data.frame with 16,674 rows of 3 variables:
\describe{
  \item{SMILES}{SMILES, in string of characters}
  \item{E}{internal energy, in kJ/mol}
  \item{HOMO-LUMO gap}{HOMO-LUMO gap, in eV}
  ...
}}
\source{
\url{https://pubchem.ncbi.nlm.nih.gov/}
}
\usage{
qspr.data
}
\description{
A dataset containing SMILES of 16,674 compounds from PubChem with
associated internal energy E and HOMO-LUMO gap
}
\keyword{datasets}
