% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trainedSMI}
\alias{trainedSMI}
\title{List of 5,000 SMILES used for training the n-grams model}
\format{A character vector with 5,000 rows of 1 variable:
\describe{
  ...
}}
\source{
\url{https://pubchem.ncbi.nlm.nih.gov/}
}
\usage{
trainedSMI
}
\description{
A dataset containing SMILES of 5,000 compounds from PubChem
}
\keyword{datasets}
