% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewstr.R
\name{viewstr}
\alias{viewstr}
\title{view 2D structures from SMILES string vector}
\usage{
viewstr(smis = NULL, width = 500, height = 500, nrow = 2, ncol = 2,
  legend = "")
}
\arguments{
\item{smis}{is a SMILES strings vector to submit.}

\item{width}{of the picture}

\item{height}{of the picture}

\item{nrow}{is the number of rows for the table of molecular depictions}

\item{ncol}{is the number of columns for the table of molecular depictions}

\item{legend}{is the text to appear for each molecular depiction in the top left corner}
}
\value{
a plot of requested 2D structures.
}
\description{
View 2D structures from SMILES strings vector
}
\examples{
viewstr(c("c1ccc2ccc3c(NCCN(C)C)cc(nc3c2c1)", "c1ccc2ccc3c(NCCN(CC)CCCl)cc(nc3c2c1)",
 "c1ccc2ccc3c(NC(CC)CC)cc(nc3c2c1)", "c1ccc2ccc3c(c2c1)ncc(c3NCCNCC=CCCCC)"))

}
